Models.py updates to order and existing database
Added "__table_args__ = {'extend_existing': True}" to work with existing test sqlite3 database. Can probably be removed when starting fresh
Added "Ordered = True" option to return a consistent JSON format in the order it was serialized which should match same as CSV source header order. This is mainly for human convenience of course.